Update Leave Chat logic in Conversation Transfer + added more checks when a supervisor attempts to barge in (logic around if they are already a member of the conversation or not)

+++ plugin-flex-ts-template-v2/src/feature-library/supervisor-barge-coach/custom-components/ChatBargeButton/SupervisorChatBargeButton.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as Flex from '@twilio/flex-ui';
 import { useFlexSelector, ITask, templates } from '@twilio/flex-ui';
 import { useDispatch, useSelector } from 'react-redux';
-import { Stack } from '@twilio-paste/core/stack';
+import { Box } from '@twilio-paste/core/box';
 import { Button } from '@twilio-paste/core/button';
 import { Flex as FlexBox } from '@twilio-paste/core/flex';
 
@@ -19,27 +19,32 @@ type SupervisorChatBargeProps = {
 export const SupervisorChatBargeButton = ({ task }: SupervisorChatBargeProps) => {
   const dispatch = useDispatch();
   const [processing, setProcessing] = React.useState(false);
+  const [isChecking, setIsChecking] = React.useState(false);
   const { chatBarge } = useSelector(
     (state: AppState) => state[reduxNamespace].supervisorBargeCoach as SupervisorBargeCoachState,
   );
 
   const teamViewTaskSID = useFlexSelector((state) => state?.flex?.view?.selectedTaskInSupervisorSid) || '';
   const agentWorkerSID = useFlexSelector((state) => state?.flex?.supervisor?.stickyWorker?.worker?.sid) || '';
-  const myWorkerName = useFlexSelector((state) => state?.flex?.worker?.worker?.name) || '';
+  const myWorkerName = useFlexSelector((state) => state?.flex?.session?.identity) || '';
   const conversationSid = task?.attributes?.conversationSid || '';
   const [chatBargeStatus, setChatBargeStatus] = React.useState(false);
-
   // Storing teamViewPath to browser cache to help if a refresh happens
   // will use this in the browserRefreshHelper
   if (teamViewTaskSID && agentWorkerSID) {
     localStorage.setItem('teamViewTaskSID', teamViewTaskSID);
     localStorage.setItem('agentWorkerSID', agentWorkerSID);
   }
-  const newChatBargeState = { ...chatBarge, [teamViewTaskSID]: !chatBargeStatus };
   const bargeHandleClick = async () => {
     setProcessing(true);
+    if (!teamViewTaskSID) {
+      setProcessing(false);
+      return;
+    }
     if (chatBargeStatus) {
       await BargeCoachService.removeWorkerParticipant(conversationSid, myWorkerName);
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const { [teamViewTaskSID]: value, ...newChatBargeState } = chatBarge;
       dispatch(Actions.setBargeCoachStatus({ chatBarge: newChatBargeState }));
       setChatBargeStatus(false);
       // Storing chatBarge to browser cache to help if a refresh happens
@@ -47,6 +52,7 @@ export const SupervisorChatBargeButton = ({ task }: SupervisorChatBargeProps) =>
       localStorage.setItem('chatBarge', JSON.stringify(newChatBargeState));
     } else {
       await BargeCoachService.inviteWorkerParticipant(conversationSid, myWorkerName);
+      const newChatBargeState = { ...chatBarge, [teamViewTaskSID]: true };
       dispatch(Actions.setBargeCoachStatus({ chatBarge: newChatBargeState }));
       setChatBargeStatus(true);
       // Storing chatBarge to browser cache to help if a refresh happens
@@ -57,19 +63,57 @@ export const SupervisorChatBargeButton = ({ task }: SupervisorChatBargeProps) =>
     Flex.Actions.invokeAction('SelectTaskInSupervisor', { sid: teamViewTaskSID });
     setProcessing(false);
   };
+  React.useEffect(() => {
+    const checkSupervisorInChat = async () => {
+      if (teamViewTaskSID in chatBarge) {
+        setChatBargeStatus(chatBarge[teamViewTaskSID]);
+      } else {
+        setIsChecking(true);
+        try {
+          const supervisorInChat = await BargeCoachService.workerPartOfConversation(conversationSid, myWorkerName);
+          // If supervisor is in the chat, update the barge state.
+          if (supervisorInChat) {
+            if (!teamViewTaskSID) {
+              setIsChecking(false);
+              return;
+            }
+            const newChatBargeState = { ...chatBarge, [teamViewTaskSID]: true };
+            dispatch(Actions.setBargeCoachStatus({ chatBarge: newChatBargeState }));
+            setChatBargeStatus(true);
+            localStorage.setItem('chatBarge', JSON.stringify(newChatBargeState));
+          } else {
+            setChatBargeStatus(false);
+          }
+        } catch (err) {
+          console.error('Failed to fetch chat participants:', err);
+          setChatBargeStatus(false);
+        }
+        setIsChecking(false);
+      }
+    };
+    checkSupervisorInChat();
+  }, [teamViewTaskSID, chatBarge]);
 
   React.useEffect(() => {
-    if (teamViewTaskSID in chatBarge) {
-      setChatBargeStatus(chatBarge[teamViewTaskSID]);
-    } else {
+    if (chatBargeStatus && (task.status === 'wrapping' || task.status === 'completed')) {
+      console.log('Attempting to remove worker from conversation');
+      BargeCoachService.removeWorkerParticipant(conversationSid, myWorkerName);
       setChatBargeStatus(false);
     }
-  }, [teamViewTaskSID, chatBarge]);
+  }, [task]);
+
+  const isLiveConversation = (task: ITask): boolean =>
+    task !== null && task.status !== 'completed' && task.status !== 'wrapping';
 
   return (
     <FlexBox hAlignContent="left" vertical>
-      <Stack orientation="horizontal" spacing="space30" element="BARGE_COACH_BUTTON_BOX">
-        <Button variant="primary" size="small" onClick={bargeHandleClick} disabled={processing}>
+      <Box padding="space10" element="BARGE_COACH_BUTTON_BOX">
+        <Button
+          variant="primary"
+          size="small"
+          onClick={bargeHandleClick}
+          disabled={processing || isChecking || !isLiveConversation(task)}
+        >
           {processing
             ? chatBargeStatus
               ? templates[StringTemplates.Leaving]()
@@ -78,7 +122,7 @@ export const SupervisorChatBargeButton = ({ task }: SupervisorChatBargeProps) =>
             ? templates[StringTemplates.Leave]()
             : templates[StringTemplates.Join]()}
         </Button>
-      </Stack>
+      </Box>
     </FlexBox>
   );
 };
